move calculation and stream in separate class
